Rewrite git recipe (notes on rewriting history); fix anchor bug
The anchor hack to allow for the fixed header makes the h* elements
extra large and extend 70ish pixels above their box. The first h3
covered the bottom items in the Gir Recipe Book table of contents.

Now it's placed below (z-index -1) so it doesn't prevent the table of
contents from working