Update blog post summary tile so it doesn't flash on hover
Previously, when hovered, it animates left and up a little bit. If you
hover from the bottom, it animates away from your cursor -- so it loses
the hover state, so it moves back to it's original position, so it's
under your cursor, so it animates away, etc. The effect is a
twitchy-flashy thing that does not look good.