Add Mastodon link to latest blog post; Update blog post index animation
Now that I have >20 blog posts, the last few were not animating in. You
only notice it if you're scrolled down to the bottom of the page and
refresh, but it was bugging me. I have another 17 blog posts to write
before I have to deal with this problem again.